ref(app-starts): Move module to performance dir
Move the App Starts module to the performance directory for long term
maintability
