+++ static/app/views/performance/appStarts/screenSummary/startTypeSelector.tsx
