+++ static/app/views/performance/appStarts/screens/averageComparisonChart.tsx
@@ -11,6 +11,8 @@ import {decodeScalar} from 'sentry/utils/queryString';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import {useLocation} from 'sentry/utils/useLocation';
 import usePageFilters from 'sentry/utils/usePageFilters';
+import {MAX_CHART_RELEASE_CHARS} from 'sentry/views/performance/appStarts/screens';
+import {COLD_START_TYPE} from 'sentry/views/performance/appStarts/screenSummary/startTypeSelector';
 import {YAxis, YAXIS_COLUMNS} from 'sentry/views/performance/screenload/screens';
 import {ScreensBarChart} from 'sentry/views/performance/screenload/screens/screenBarChart';
 import {useTableQuery} from 'sentry/views/performance/screenload/screens/screensTable';
@@ -19,8 +21,6 @@ import {useReleaseSelection} from 'sentry/views/starfish/queries/useReleases';
 import {SpanMetricsField} from 'sentry/views/starfish/types';
 import {formatVersionAndCenterTruncate} from 'sentry/views/starfish/utils/centerTruncate';
 import {appendReleaseFilters} from 'sentry/views/starfish/utils/releaseComparison';
-import {MAX_CHART_RELEASE_CHARS} from 'sentry/views/starfish/views/appStartup';
-import {COLD_START_TYPE} from 'sentry/views/starfish/views/appStartup/screenSummary/startTypeSelector';
 
 interface Props {
   chartHeight?: number;
